+++ Src-datafile/DataFileIMAP.cpp
@@ -132,7 +132,7 @@ ShankMap* DataFileIMAP::shankMap( bool forExport ) const
         GeomMap    *G = geomMap( false );
 
         if( G ) {
-            for( int ie = 0, ne = G->e.size(); ie < ne; ++ie )
+            for( int ie = 0, ne = (int)G->e.size(); ie < ne; ++ie )
                 shankMap->e[ie].u = G->e[ie].u;
             delete G;
         }
@@ -143,7 +143,7 @@ ShankMap* DataFileIMAP::shankMap( bool forExport ) const
 
                 S->fromString( it.value().toString() );
 
-                for( int ie = 0, ne = S->e.size(); ie < ne; ++ie )
+                for( int ie = 0, ne = (int)S->e.size(); ie < ne; ++ie )
                     shankMap->e[ie].u = S->e[ie].u;
             }
 
@@ -392,7 +392,7 @@ GeomMap* DataFileIMAP::geomMap( bool forExport ) const
 
         ShankMap    *S = shankMap( false );
         if( S ) {
-            for( int ie = 0, ne = S->e.size(); ie < ne; ++ie )
+            for( int ie = 0, ne = (int)S->e.size(); ie < ne; ++ie )
                 geomMap->e[ie].u = S->e[ie].u;
             delete S;
         }
